requirements.txt
agno
streamlit
chromadb
python-dotenv
pytest
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0